CC-5572: Prefix account ID to image name in containers images delete
Also fix an issue with the "containers images list" command where the
final line of output was being truncated.
