+++ .changeset/four-needles-rule.md
@@ -1,0 +1,5 @@
+---
+"wrangler": patch
+---
+
+Fix unauthorized errors on "containers images delete".
