+++ packages/wrangler/src/__tests__/cloudchamber/curl.test.ts
@@ -33,7 +33,7 @@ describe("cloudchamber curl", () => {
 		expect(helpStd.out).toMatchInlineSnapshot(`
 			"wrangler cloudchamber curl <path>
 
-			send a request to an arbitrary Cloudchamber endpoint
+			Send a request to an arbitrary Cloudchamber endpoint
 
 			POSITIONALS
 			  path  [string] [required] [default: \\"/\\"]
