+++ packages/wrangler/src/__tests__/cloudchamber/images.test.ts
@@ -1,3 +1,4 @@
+import { getCloudflareContainerRegistry } from "@cloudflare/containers-shared";
 import { http, HttpResponse } from "msw";
 import patchConsole from "patch-console";
 import { vi } from "vitest";
@@ -112,8 +113,7 @@ describe("cloudchamber image", () => {
 			http.delete(
 				"*/registries/:domain",
 				async ({ params }) => {
-					const domain = String(params["domain"]);
-					expect(domain === "docker.io");
+					expect(params.domain).toEqual("docker.io");
 					return HttpResponse.json({});
 				},
 				{ once: true }
@@ -166,6 +166,8 @@ describe("cloudchamber image list", () => {
 	const std = mockConsoleMethods();
 	const { setIsTTY } = useMockIsTTY();
 
+	const REGISTRY = getCloudflareContainerRegistry();
+
 	mockAccountId();
 	mockApiToken();
 	beforeEach(mockAccount);
@@ -183,7 +185,7 @@ describe("cloudchamber image list", () => {
 		expect(std.out).toMatchInlineSnapshot(`
 			"wrangler cloudchamber images list
 
-			perform operations on images in your Cloudflare managed registry
+			List images in the Cloudflare managed registry
 
 			GLOBAL FLAGS
 			  -c, --config   Path to Wrangler configuration file  [string]
@@ -209,11 +211,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
@@ -254,11 +255,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
@@ -297,11 +297,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
@@ -342,11 +341,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
@@ -405,11 +403,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
@@ -466,11 +463,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
@@ -486,30 +482,33 @@ describe("cloudchamber image list", () => {
 					tags: t,
 				});
 			}),
-			http.head("*/v2/:image/manifests/:tag", async ({ params }) => {
-				const image = String(params["image"]);
-				expect(image === "one");
-				const tag = String(params["tag"]);
-				expect(tag === "hundred");
+			http.head("*/v2/:accountId/:image/manifests/:tag", async ({ params }) => {
+				expect(params.accountId).toEqual("some-account-id");
+				expect(params.image).toEqual("one");
+				expect(params.tag).toEqual("hundred");
 				return new HttpResponse("", {
 					status: 200,
 					headers: { "Docker-Content-Digest": "some-digest" },
 				});
 			}),
-			http.delete("*/v2/:image/manifests/:tag", async ({ params }) => {
-				const image = String(params["image"]);
-				expect(image === "one");
-				const tag = String(params["tag"]);
-				expect(tag === "hundred");
-				return new HttpResponse("", { status: 200 });
-			}),
+			http.delete(
+				"*/v2/:accountId/:image/manifests/:tag",
+				async ({ params }) => {
+					expect(params.accountId).toEqual("some-account-id");
+					expect(params.image).toEqual("one");
+					expect(params.tag).toEqual("hundred");
+					return new HttpResponse("", { status: 200 });
+				}
+			),
 			http.put("*/v2/gc/layers", async () => {
 				return new HttpResponse("", { status: 200 });
 			})
 		);
 		await runWrangler("cloudchamber images delete one:hundred");
 		expect(std.err).toMatchInlineSnapshot(`""`);
-		expect(std.out).toMatchInlineSnapshot(`"Deleted tag: one:hundred"`);
+		expect(std.out).toMatchInlineSnapshot(
+			`"Deleted one:hundred (some-digest)"`
+		);
 	});
 
 	it("should error when provided a repo without a tag", async () => {
@@ -523,11 +522,10 @@ describe("cloudchamber image list", () => {
 
 		msw.use(
 			http.post("*/registries/:domain/credentials", async ({ params }) => {
-				const domain = String(params["domain"]);
-				expect(domain === "docker.io");
+				expect(params.domain).toEqual(REGISTRY);
 				return HttpResponse.json({
 					account_id: "1234",
-					registry_host: "docker.io",
+					registry_host: REGISTRY,
 					username: "foo",
 					password: "bar",
 				});
