+++ packages/wrangler/src/__tests__/containers/list.test.ts
@@ -27,7 +27,7 @@ describe("containers list", () => {
 		expect(std.out).toMatchInlineSnapshot(`
 			"wrangler containers list
 
-			list containers
+			List containers
 
 			GLOBAL FLAGS
 			  -c, --config   Path to Wrangler configuration file  [string]
